requirements.txt
tgb
aiohttp
lxml
asyncpg
pendulum
click
ujson
pyyaml
